+++ core-rust/mesh-api-server/src/mesh_api/handlers/mod.rs
@@ -1,13 +1,11 @@
 mod account_balance;
 mod block;
-mod block_transaction;
 mod network_list;
 mod network_options;
 mod network_status;
 
 pub(crate) use account_balance::*;
 pub(crate) use block::*;
-pub(crate) use block_transaction::*;
 pub(crate) use network_list::*;
 pub(crate) use network_options::*;
 pub(crate) use network_status::*;
